+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/IllegalTypeCheckTest.java
@@ -45,7 +45,9 @@ public class IllegalTypeCheckTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "coding" + File.separator + filename);
+                + "coding" + File.separator
+                + "illegaltype" + File.separator
+                + filename);
     }
 
     @Test
@@ -85,7 +87,8 @@ public class IllegalTypeCheckTest extends BaseCheckTestSupport {
         final String[] expected = {
             "6:13: " + getCheckMessage(MSG_KEY, "AbstractClass"),
             "9:13: " + getCheckMessage(MSG_KEY,
-                "com.puppycrawl.tools.checkstyle.checks.coding.InputIllegalType.AbstractClass"),
+                "com.puppycrawl.tools.checkstyle.checks.coding.illegaltype."
+                    + "InputIllegalType.AbstractClass"),
             "16:13: " + getCheckMessage(MSG_KEY, "java.util.TreeSet"),
         };
 
@@ -111,7 +114,8 @@ public class IllegalTypeCheckTest extends BaseCheckTestSupport {
 
         final String[] expected = {
             "9:13: " + getCheckMessage(MSG_KEY,
-                "com.puppycrawl.tools.checkstyle.checks.coding.InputIllegalType.AbstractClass"),
+                "com.puppycrawl.tools.checkstyle.checks.coding.illegaltype."
+                    + "InputIllegalType.AbstractClass"),
             "16:13: " + getCheckMessage(MSG_KEY, "java.util.TreeSet"),
             "17:13: " + getCheckMessage(MSG_KEY, "TreeSet"),
         };
@@ -135,10 +139,10 @@ public class IllegalTypeCheckTest extends BaseCheckTestSupport {
     @Test
     public void testSameFileNameGeneral() throws Exception {
         checkConfig.addAttribute("illegalClassNames",
-            "List, InputGregorianCalendar, java.io.File, ArrayList");
+            "List, InputIllegalTypeGregorianCalendar, java.io.File, ArrayList");
         final String[] expected = {
-            "10:5: " + getCheckMessage(MSG_KEY, "InputGregorianCalendar"),
-            "16:23: " + getCheckMessage(MSG_KEY, "InputGregorianCalendar"),
+            "10:5: " + getCheckMessage(MSG_KEY, "InputIllegalTypeGregorianCalendar"),
+            "16:23: " + getCheckMessage(MSG_KEY, "InputIllegalTypeGregorianCalendar"),
             "24:9: " + getCheckMessage(MSG_KEY, "List"),
             "25:9: " + getCheckMessage(MSG_KEY, "java.io.File"),
             "27:5: " + getCheckMessage(MSG_KEY, "java.util.List"),
@@ -179,12 +183,12 @@ public class IllegalTypeCheckTest extends BaseCheckTestSupport {
         final String[] expected = {
             "6:13: " + getCheckMessage(MSG_KEY, "AbstractClass"),
             "9:13: " + getCheckMessage(MSG_KEY,
-                "com.puppycrawl.tools.checkstyle.checks.coding."
+                "com.puppycrawl.tools.checkstyle.checks.coding.illegaltype."
                     + "InputIllegalTypeMemberModifiers.AbstractClass"),
             "16:13: " + getCheckMessage(MSG_KEY, "java.util.TreeSet"),
             "17:13: " + getCheckMessage(MSG_KEY, "TreeSet"),
             "23:15: " + getCheckMessage(MSG_KEY,
-                "com.puppycrawl.tools.checkstyle.checks.coding."
+                "com.puppycrawl.tools.checkstyle.checks.coding.illegaltype."
                     + "InputIllegalTypeMemberModifiers.AbstractClass"),
             "25:25: " + getCheckMessage(MSG_KEY, "java.util.TreeSet"),
             "33:15: " + getCheckMessage(MSG_KEY, "AbstractClass"),
