+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/illegaltype/InputIllegalTypeGregorianCalendar.java
@@ -1,0 +1,8 @@
+package com.puppycrawl.tools.checkstyle.checks.coding.illegaltype;
+
+public class InputIllegalTypeGregorianCalendar
+{
+    class SubCalendar {
+        
+    }
+}
