+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/illegaltype/InputIllegalTypeMemberModifiers.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.illegaltype;
 import java.util.TreeSet;
 import java.util.Hashtable;
 //configuration: default
@@ -6,8 +6,8 @@ public class InputIllegalTypeMemberModifiers {
     private AbstractClass a = null; //WARNING
     private NotAnAbstractClass b = null; /*another comment*/
 
-    private com.puppycrawl.tools.checkstyle.checks.coding.InputIllegalTypeMemberModifiers.AbstractClass c = null; //WARNING
-    private com.puppycrawl.tools.checkstyle.checks.coding.InputIllegalTypeMemberModifiers.NotAnAbstractClass d = null;
+    private com.puppycrawl.tools.checkstyle.checks.coding.illegaltype.InputIllegalTypeMemberModifiers.AbstractClass c = null; //WARNING
+    private com.puppycrawl.tools.checkstyle.checks.coding.illegaltype.InputIllegalTypeMemberModifiers.NotAnAbstractClass d = null;
 
     private abstract class AbstractClass {/*one more comment*/}
 
@@ -20,7 +20,7 @@ public class InputIllegalTypeMemberModifiers {
     }
     
     //WARNING if memberModifiers is set and contains TokenTypes.LITERAL_PROTECTED
-    protected com.puppycrawl.tools.checkstyle.checks.coding.InputIllegalTypeMemberModifiers.AbstractClass c1 = null;
+    protected com.puppycrawl.tools.checkstyle.checks.coding.illegaltype.InputIllegalTypeMemberModifiers.AbstractClass c1 = null;
     //NO WARNING if memberModifiers is set and does not contain TokenTypes.LITERAL_PUBLIC
     public final static java.util.TreeSet<Object> table3() { return null; }
