+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/illegaltype/InputIllegalTypeStaticImports.java
@@ -1,8 +1,8 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.illegaltype;
 
 import static com.puppycrawl.tools.checkstyle.utils.CheckUtils.isElseIf;
 import static com.puppycrawl.tools.checkstyle.utils.CheckUtils.*;
-import static com.puppycrawl.tools.checkstyle.checks.coding.InputIllegalType.SomeStaticClass;
+import static com.puppycrawl.tools.checkstyle.checks.coding.illegaltype.InputIllegalType.SomeStaticClass;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 //configuration: "illegalClassNames": SomeStaticClass
 public class InputIllegalTypeStaticImports
